1、删除自提点 2、创建资产

+++ src/main/java/com/hdu/lease/pojo/request/CreateAssertRequest.java
@@ -35,7 +35,7 @@ public class CreateAssertRequest {
     /**
      * 价值
      */
-    private Double value;
+    private int value;
 
     /**
      * 自提点列表
